Improve performance of RGraph library - Line graph
RGraph currently uses object oriented functions which has performance issues
as the number of objects increases. This is due to the duplication of the
class methods in each of the object.

By using prototype based OOP, we can save the overhead of copying the class
methods in each object.

This fixes #2 for the line graph. As we go along, we should optimize the remaining
classes.